feat(client): client call event reporting - move to reporting
