+++ packages/client/src/reporting/__tests__/ClientEventReporter.test.ts
