+++ packages/client/src/reporting/index.ts
@@ -1,0 +1,1 @@
+export * from './ClientEventReporter';
