+++ packages/client/src/stats/index.ts
@@ -1,5 +1,4 @@
 export * from './CallStateStatsReporter';
-export * from './ClientEventReporter';
 export * from './SfuStatsReporter';
 export * from './types';
 export * from './utils';
